refactor: rename flip_logging to log_config
Rename the shared logging library from flip_logging to log_config
for clarity. Update all imports, compose volume mounts, CI workflow
paths, gitignores, and documentation.
