+++ trust/alloy-config.alloy
@@ -1,6 +1,6 @@
 // Grafana Alloy configuration -- replaces Promtail for Docker log collection.
 // Discovers containers via the Docker socket, parses JSON logs from
-// flip_logging, and pushes to Loki.
+// log_config, and pushes to Loki.
 
 discovery.docker "docker" {
   host             = "unix:///var/run/docker.sock"
@@ -35,7 +35,7 @@ loki.source.docker "docker" {
   refresh_interval = "5s"
 }
 
-// Parse JSON log lines from flip_logging and promote fields to Loki labels.
+// Parse JSON log lines from log_config and promote fields to Loki labels.
 loki.process "json_parse" {
   forward_to = [loki.write.loki.receiver]
