+++ trust/imaging-api/imaging_api/utils/logger.py
@@ -10,7 +10,7 @@
 # limitations under the License.
 #
 
-from flip_logging import configure_logging, get_logger
+from log_config import configure_logging, get_logger
 
 from imaging_api.config import get_settings
