+++ trust/log_config/formatter.py
